Rename qiskit/tools/events/_pubsub
